return all infra

+++ src/main/java/com/siloenix/socket/errors/ErrorCode.java
@@ -1,5 +1,5 @@
 package com.siloenix.socket.errors;
 
 public interface ErrorCode {
-    int code();
+    byte code();
 }
